PInvoke workaround to read Moonlight CLI output

+++ PlayTogetherMain/PlayTogetherMod.cs
@@ -19,6 +19,8 @@ using System.Linq;
 using BTKUILib.UIObjects.Components;
 using System.Text.RegularExpressions;
 using System.Security.Cryptography;
+using System.Runtime.InteropServices;
+using System.Text;
 
 namespace PlayTogetherMod
 {
@@ -30,16 +32,15 @@ namespace PlayTogetherMod
     public class Sunshine
     {
         public const string EXE_PATH = SharedVars.RESOURCE_FOLDER + @"\Sunshine\sunshine.exe";
-        public const string SETTINGS_PATH = SharedVars.RESOURCE_FOLDER + @"\Sunshine\config\sunshine.conf"; //Multicast may or may not be necessary for multiple clients, idk.
-        public const string APPSCONF_PATH = SharedVars.RESOURCE_FOLDER + @"\Sunshine\assets\apps.json";
-        public const string APPSCONF_PATH2 = SharedVars.RESOURCE_FOLDER + @"\Sunshine\config\apps.json";
-        bool firstLaunch = true;
-        public Process firststartprocess;
+        public const string SETTINGS_PATH = SharedVars.RESOURCE_FOLDER + @"\Sunshine\config\sunshine.conf"; // Multicast may or may not be necessary for multiple clients, idk.
+        public const string APPSDEFCONF_PATH = SharedVars.RESOURCE_FOLDER + @"\Sunshine\assets\apps.json";
+        public const string APPSCONF_DIR = SharedVars.RESOURCE_FOLDER + @"\Sunshine\config";
+        public const string APPSCONF_PATH = APPSCONF_DIR + @"\apps.json";
         public Process normalprocess;
         StreamWriter streamWriter;
-        //Apps file is only read on launch so sunshine needs to restart on edit. Probably the same for other confs.
-        //channels=3 (3 distinct streams) //3 for 4 players
-        //Configuration variables can be overwritten on the command line: "name=value" --> it can be usefull to set min_log_level=debug without modifying the configuration file
+        // Apps file is only read on launch so sunshine needs to restart on edit. Probably the same for other confs.
+        // channels=3 (3 distinct streams) //3 for 4 players
+        // Configuration variables can be overwritten on the command line: "name=value" --> it can be usefull to set min_log_level=debug without modifying the configuration file
         ProcessStartInfo firststartinfo = new ProcessStartInfo {
             FileName = EXE_PATH,
             WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Sunshine",
@@ -52,12 +53,12 @@ namespace PlayTogetherMod
         {
             FileName = EXE_PATH,
             WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Sunshine",
-            Arguments = "-p -0 -1",
+            Arguments = "-p -0",
             UseShellExecute = false,
             RedirectStandardInput = true,
-            RedirectStandardOutput = false //True somehow makes the process unable to receive our inputs
+            RedirectStandardOutput = false
         };
-        //Make a warning for users to make sure apps they add are always launched fullscreen for privacy reasons. Atl-tabbing risky, theres probably a way to prevent desktop from streaming. (apps.json?)
+        // Make a warning for users to make sure apps they add are always launched fullscreen for privacy reasons. Atl-tabbing risky, theres probably a way to prevent desktop from streaming. (apps.json?)
 
         public void WritePin(string pin)
         {
@@ -65,20 +66,17 @@ namespace PlayTogetherMod
             //streamWriter.Close();
         }
 
+        public void FlushConfigs()
+        {
+            if(Directory.Exists(APPSCONF_DIR))
+                Directory.Delete(APPSCONF_DIR, true);
+        }
+
         public void Run(string appPath)
         {
-            if (firstLaunch)
-            {
-                firstLaunch = false;
-                firststartprocess = Process.Start(firststartinfo);
-                if (!firststartprocess.HasExited)
-                {
-                    firststartprocess.WaitForExit();
-                }
-            }
+            FlushConfigs();
             string appstr = @"{""name"":""" + Path.GetFileNameWithoutExtension(appPath) + @""",""cmd"":""" + Regex.Replace(appPath, @"\\|/", @"\\") + @""",""auto-detach"":""true"",""wait-all"":""true"",""image-path"":""steam.png""}";
-            File.WriteAllText(APPSCONF_PATH, @"{""env"":{},""apps"":[" + appstr + @"]}"); //Temporary lazy fix for json issues
-            File.WriteAllText(APPSCONF_PATH2, @"{""env"":{},""apps"":[" + appstr + @"]}"); //Temporary lazy fix for json issues
+            File.WriteAllText(APPSDEFCONF_PATH, @"{""env"":{},""apps"":[" + appstr + @"]}"); //Temporary lazy fix for json issues
             normalprocess = Process.Start(normalstartinfo);
             streamWriter = normalprocess.StandardInput;
         }
@@ -97,15 +95,77 @@ namespace PlayTogetherMod
     {
         public const string EXE_PATH = SharedVars.RESOURCE_FOLDER + @"\Moonlight\Moonlight.exe";
         public const string INI_PATH = SharedVars.RESOURCE_FOLDER + @"\Moonlight\Moonlight Game Streaming Project\Moonlight.ini";
-        public Process normalprocess;
+        public Process sessionProc;
+        string _lobbyCode = "";
+
+        // Import the required WinAPI functions to Pinvoke
+        // This workaround is needed for reading Moonlight CLI's outputs
+        //  as it redirects it's stdout asynchronously to handle it's log writes right before shutdown.
+        //  This prevents us from capturing the info we need unless we directly read the console output buffer after the fact.
+        //  (which is what the PInvoke is for)
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern IntPtr GetStdHandle(int nStdHandle);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern bool ReadConsoleOutputCharacter(IntPtr hConsoleOutput, [Out] StringBuilder lpCharacter, uint nLength, COORD dwReadCoord, out uint lpNumberOfCharsRead);
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct COORD
+        {
+            public short X;
+            public short Y;
+        }
+
+        const int STD_OUTPUT_HANDLE = -11;
+
+        // Function to read the console output
+        static string? ReadConsoleOutput()
+        {
+            // Get the handle to the standard output console buffer
+            IntPtr hConsoleOutput = GetStdHandle(STD_OUTPUT_HANDLE);
+            if (hConsoleOutput == IntPtr.Zero || hConsoleOutput == (IntPtr)(-1))
+            {
+                Console.WriteLine("Error: Failed to get console output handle.");
+                return null;
+            }
+
+            // Define the buffer size
+            const int bufferSize = 4096;
+
+            // Create a buffer to store the console output
+            StringBuilder buffer = new StringBuilder(bufferSize);
+
+            // Read the console output into the buffer
+            uint charsRead;
+            COORD readCoord = new COORD { X = 0, Y = 0 };
+            if (!ReadConsoleOutputCharacter(
+                hConsoleOutput,
+                buffer,
+                (uint)bufferSize,
+                readCoord,
+                out charsRead))
+            {
+                //Failed to read console output
+                return null;
+            }
+
+            // Trim the StringBuilder to the actual number of characters read
+            buffer.Length = (int)charsRead;
+
+            // Return the console output as a string
+            return buffer.ToString();
+        }
 
         public void PairWithHost(string lobbyCode, string pairPin)
         {
+            //_lobbyCode = lobbyCode; //uncomment after debugging
+            _lobbyCode = "2817135310";
+            pairPin = "1234"; //remove after debugging
             Process pairprocess = Process.Start(new ProcessStartInfo()
             {
                 FileName = EXE_PATH,
                 WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Moonlight",
-                Arguments = $"pair {LobbyCodeHandler.LobbyCodeToIP(lobbyCode)} --pin {pairPin}",
+                Arguments = $"pair {LobbyCodeHandler.LobbyCodeToIP(_lobbyCode)} --pin {pairPin}",
                 UseShellExecute = false,
                 RedirectStandardInput = true,
                 RedirectStandardOutput = true
@@ -114,20 +174,75 @@ namespace PlayTogetherMod
 
         public void ClearInitFile() //Intended to prevent endless accumulation of expired clients
         {
-            File.Delete(INI_PATH); //Not sure about the implications of deleting the [gcmapping] field yet. May need granular control over this file depending on results.
+            if(File.Exists(INI_PATH))
+                File.Delete(INI_PATH); //Not sure about the implications of deleting the [gcmapping] field yet. May need granular control over this file depending on results.
         }
 
-        public void Run()
+        private string? GetHostAppTitle(string host)
         {
-            //Would normally use the list command but it doesn't return anything when using the portable version of the apps. (could be either or both)
-            normalprocess = Process.Start(new ProcessStartInfo()
+            using (var dummyprocess = Process.Start(new ProcessStartInfo()
             {
                 FileName = EXE_PATH,
                 WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Moonlight",
+                Arguments = $"pair {host}",
+                UseShellExecute = false,
+                RedirectStandardInput = true,
+                RedirectStandardOutput = true
+            })){
+                dummyprocess.WaitForExit(); // Might never happen until user clicks.. Manually stopping might be necessary
+            };
+            //Stop(dummyprocess);
+            using (var listprocess = Process.Start(new ProcessStartInfo()
+            {
+                FileName = EXE_PATH,
+                WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Moonlight",
+                Arguments = $"list {host}",
+                UseShellExecute = false,
+                RedirectStandardInput = true,
+                RedirectStandardOutput = true
+            }))
+            {
+                if (listprocess == null)
+                {
+                    // Error: Failed to start Moonlight process.
+                    return null;
+                }
+                listprocess.WaitForExit();
+                return ReadConsoleOutput(); //Returns null 
+            };
+        }
+
+        public void StartSession(string host, string? hostAppTitle)
+        {
+            if((hostAppTitle == null) || (hostAppTitle == ""))
+            {
+                return;
+            }
+            sessionProc = Process.Start(new ProcessStartInfo()
+            {
+                FileName = EXE_PATH,
+                WorkingDirectory = SharedVars.RESOURCE_FOLDER + @"\Moonlight",
+                Arguments = @$"stream {host} ""{hostAppTitle}""",
                 UseShellExecute = false,
                 RedirectStandardInput = true,
                 RedirectStandardOutput = true
             });
+            //sessionProc.WaitForExit();
+        }
+
+        public void Run()
+        {
+            var host = LobbyCodeHandler.LobbyCodeToIP(_lobbyCode);
+            StartSession(host, GetHostAppTitle(host));
+        }
+
+        public void Stop(Process processToStop)
+        {
+            if (!processToStop.HasExited)
+            {
+                processToStop.Kill();
+                processToStop.WaitForExit();
+            }
         }
     }
 
@@ -227,7 +342,7 @@ namespace PlayTogetherMod
             var sendPinButton = sendPinCat.AddButton("", "", "Validate a friend's PIN");
             sendPinButton.OnPress += () =>
             {
-                _sunshine.WritePin(pinInputs);
+                _sunshine.WritePin("1234");// _sunshine.WritePin(pinInputs); //Uncomment after debug
                 pinNumpad.Disabled = false;
                 pinInputs = "";
                 sendPinButton.ButtonText = pinInputs;
@@ -260,21 +375,30 @@ namespace PlayTogetherMod
             }
 
             var pairPage = clientCat.AddPage("Connect to Lobby", "", "Pair yourself with a Host", "CVRPlayTogether");
-            var joinButton = clientCat.AddButton("Join Game", "", "Attempt to join game");
-            joinButton.OnPress += () => {
-                _moonlight.Run();
+            var joinToggle = clientCat.AddToggle("Join Game", "Attempt to join game", false);
+            joinToggle.OnValueUpdated += b =>
+            {
+                if(b == true)
+                {
+                    _moonlight.Run();
+                    LoggerInstance.Msg($"Survived Moonlight Run");
+                }
+                else
+                {
+                    //_moonlight.Stop();
+                }
             };
             var lobbyNumpad = pairPage.AddCategory("Enter lobby code..");
             var confirmCat = pairPage.AddCategory("Connect to lobby");
             var connectButton = confirmCat.AddButton("", "", "Attempt connection with Host");
             connectButton.OnPress += () =>
             {
-                var pairingPin = GeneratePairingPin();
+                var pairingPin = "1234"; //GeneratePairingPin(); //uncomment after debug
                 QuickMenuAPI.ShowNotice("Pairing Pin", $"The host must enter this pin to approve your connection: {pairingPin}", null);
                 _moonlight.PairWithHost(targetLobbyCode, pairingPin);
-                lobbyNumpad.Disabled = false;
-                targetLobbyCode = "";
-                connectButton.ButtonText = targetLobbyCode;
+                //lobbyNumpad.Disabled = false;
+                //targetLobbyCode = "";
+                //connectButton.ButtonText = targetLobbyCode;
             };
             var clearLobbyCodeButton = confirmCat.AddButton("Clear", "", "Clear Lobby code");
             clearLobbyCodeButton.OnPress += () =>
